+++ src/Models/BuffTreeModel.cpp
@@ -21,7 +21,7 @@ BuffTreeModel::BuffTreeModel(QObject* parent) : BaseTreeModel(parent) {
 		}
 	}
 
-	for (int i = 0; i < buff_slk.rows(); i++) {
+	for (uint64_t i = 0; i < buff_slk.rows(); i++) {
 		const std::string& id = buff_slk.index_to_row.at(i);
 
 		BaseTreeItem* parent_item = getFolderParent(id);
